Changing the layout of service form to simplify runtime selection
