+++ plugins/services/src/js/constants/DefaultApp.js
@@ -1,5 +1,11 @@
+import ContainerConstants from './ContainerConstants';
+
+const {DOCKER} = ContainerConstants.type;
+
 const DEFAULT_APP_RESOURCES = {cpus: 1, mem: 128};
-const DEFAULT_APP_SPEC = Object.assign({instances: 1}, DEFAULT_APP_RESOURCES);
+const DEFAULT_APP_CONTAINER = {container: {type: DOCKER}};
+const DEFAULT_APP_SPEC = Object.assign({instances: 1},
+  DEFAULT_APP_CONTAINER, DEFAULT_APP_RESOURCES);
 
 module.exports = {
   DEFAULT_APP_RESOURCES,
